Add files via upload
There are many models for **clustering** out there. In this notebook, we will explore and compare some of the most renowned algorithms, using various metrics introduced in the class. Later, we will determine the optimal number of clusters by employing the Silhouette coefficient.

The clustering algorithms that we are going to use are as follows:
- K-Means
- Agglomerative (Single Linkage, Complete Linkage)
- DBSCAN

The measures that we evaluate the clustering algorithms' performance are as follows:
- Cluster Purity
- Rand Index
- Silhouette coefficient


In this notebook we practice clustering with 2 example to show the differences of clustering algorithms:
- A random generated dataset (non-elliptical)
- A random generated dataset (elliptical)